requirements.txt
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
typer>=0.9.0
rich>=13.0.0
jinja2>=3.1.0
playwright>=1.40.0
requests>=2.31.0
